+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Others
 
 - cli: v2.2.1 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/cli/CHANGELOG.md))
+- cloud: v2.2.0 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/cloud/CHANGELOG.md))
 
 <a name="4.3.0"></a>